Split PalveluvaylaServlet into HSL-only HslServlet
DVV moved to SuomiFiServlet (mTLS/JSON); only HSL still needs
palveluväylä. Rename PalveluvaylaServlet -> HslServlet (URL unchanged,
HSL has it hardcoded) and RequiresSuomiFiOrHsl -> RequiresHsl, now
enforcing hasHSLAccess itself like RequiresSuomiFi instead of splitting
auth across trait and servlet. Drop vainPalveluvayla, now unused.
